+++ src/game/client/hl/hl_weapons.cpp
@@ -227,11 +227,11 @@ CBaseEntity::FireBulletsPlayer
 Only produces random numbers to match the server ones.
 =====================
 */
-Vector CBaseEntity::FireBulletsPlayer ( ULONG cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
+Vector CBaseEntity::FireBulletsPlayer (uint32 cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
 {
 	float x, y, z;
 
-	for ( ULONG iShot = 1; iShot <= cShots; iShot++ )
+	for (uint32 iShot = 1; iShot <= cShots; iShot++ )
 	{
 		if ( pevAttacker == NULL )
 		{
